Bump cats-effect to 1.3.1 (closes #105)

+++ modules/id-emitter/src/main/scala/com/snowplowanalytics/snowplow/scalatracker/emitters/id/RequestProcessor.scala
@@ -52,7 +52,7 @@ class RequestProcessor {
         override def run(): Unit = queue.put(event)
       }
       val delay = getDelay(event.attempt)
-      timer.schedule(task, delay)
+      timer.schedule(task, delay.toLong)
       true
     }
